+++ src/main/java/me/exyin/jobQuests/model/rewards/interfaces/Reward.java
@@ -4,5 +4,5 @@ import java.util.UUID;
 
 public interface Reward {
     String getDescription(double quantity);
-    void giveReward(UUID uuid, String jobId, int quantity);
+    void giveReward(UUID uuid, String jobId, double quantity);
 }
